# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SIAKAD - Sistem Informasi Akademik SMP Negeri 2 Rambang
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /siswa:
    get:
      operationId: listSiswa
      tags: [siswa]
      summary: Daftar semua siswa
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: kelas_id
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Daftar siswa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiswaListResponse"
    post:
      operationId: createSiswa
      tags: [siswa]
      summary: Tambah siswa baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSiswaBody"
      responses:
        "201":
          description: Siswa berhasil ditambahkan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Siswa"
  /siswa/{id}:
    get:
      operationId: getSiswa
      tags: [siswa]
      summary: Detail siswa
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Data siswa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Siswa"
    put:
      operationId: updateSiswa
      tags: [siswa]
      summary: Update data siswa
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSiswaBody"
      responses:
        "200":
          description: Data siswa diperbarui
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Siswa"
    delete:
      operationId: deleteSiswa
      tags: [siswa]
      summary: Hapus siswa
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Siswa berhasil dihapus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # GURU
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Siswa:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
          nullable: true
        nis:
          type: string
        nisn:
          type: string
        nama:
          type: string
        kelas_id:
          type: string
          nullable: true
        kelas:
          $ref: "#/components/schemas/Kelas"
          nullable: true
        jenis_kelamin:
          type: string
          enum: [L, P]
        tanggal_lahir:
          type: string
        alamat:
          type: string
          nullable: true
        no_hp_ortu:
          type: string
          nullable: true
        foto_url:
          type: string
          nullable: true
        created_at:
          type: string
      required:
        - id
        - nis
        - nisn
        - nama
        - jenis_kelamin
        - tanggal_lahir
    SiswaListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Siswa"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CreateSiswaBody:
      type: object
      properties:
        nis:
          type: string
        nisn:
          type: string
        nama:
          type: string
        kelas_id:
          type: string
          nullable: true
        jenis_kelamin:
          type: string
          enum: [L, P]
        tanggal_lahir:
          type: string
        alamat:
          type: string
          nullable: true
        no_hp_ortu:
          type: string
          nullable: true
      required:
        - nis
        - nisn
        - nama
        - jenis_kelamin
        - tanggal_lahir
    UpdateSiswaBody:
      type: object
      properties:
        nis:
          type: string
        nisn:
          type: string
        nama:
          type: string
        kelas_id:
          type: string
          nullable: true
        jenis_kelamin:
          type: string
          enum: [L, P]
        tanggal_lahir:
          type: string
        alamat:
          type: string
          nullable: true
        no_hp_ortu:
          type: string
          nullable: true
    Guru:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
          nullable: true
        nip:
          type: string
          nullable: true
        nama:
          type: string
        mata_pelajaran_id:
          type: string
          nullable: true
        mata_pelajaran:
          $ref: "#/components/schemas/MataPelajaran"
          nullable: true
        no_hp:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        created_at:
          type: string
      required:
        - id
        - nama
...
    Kelas:
      type: object
      properties:
        id:
          type: string
        nama_kelas:
          type: string
        tingkat:
          type: integer
        wali_kelas_id:
          type: string
          nullable: true
        wali_kelas:
          $ref: "#/components/schemas/Guru"
          nullable: true
        tahun_ajaran:
          type: string
        jumlah_siswa:
          type: integer
          nullable: true
      required:
        - id
        - nama_kelas
        - tingkat
        - tahun_ajaran
...
    MataPelajaran:
      type: object
      properties:
        id:
          type: string
        nama_mapel:
          type: string
        kode_mapel:
          type: string
          nullable: true
        guru_id:
          type: string
          nullable: true
        guru:
          $ref: "#/components/schemas/Guru"
          nullable: true
      required:
        - id
        - nama_mapel
...